add option to use existing certs

+++ docker-entrypoint.sh
@@ -3,6 +3,7 @@
 set -eu
 
 echo "running docker-entrypoint.sh"
+echo "Version 2018.11.01.01"
 
 if [[ ! -d "/opt/certs" ]]; then
 	echo "/opt/certs folder is not present.  Be sure to attach a volume."
@@ -16,52 +17,55 @@ echo "-------"
 ls /opt/certs/server/
 echo "-------"
 
-if [[ ! -f "/opt/certs/server/cert.pem" ]]
+if [[ -z "${USE_EXISTING_CERTIFICATE:-}" ]]
 then
-	echo "no certificates found so regenerating them"
-
-	# make sure CertHostName and CertPassword are set
-	if [[ ! -z "${CERT_HOSTNAME_FILE:-}" ]]
+	if [[ ! -f "/opt/certs/server/cert.pem" ]]
 	then
-		echo "CERT_HOSTNAME_FILE is set so reading from $CERT_HOSTNAME_FILE"
-		CERT_HOSTNAME=$(cat $CERT_HOSTNAME_FILE)
-	fi
+		echo "no certificates found so regenerating them"
 
-	if [[ -z "${CERT_HOSTNAME:-}" ]]; then
-		echo "CERT_HOSTNAME must be set"
-    	exit 1
-	fi
+		# make sure CertHostName and CertPassword are set
+		if [[ ! -z "${CERT_HOSTNAME_FILE:-}" ]]
+		then
+			echo "CERT_HOSTNAME_FILE is set so reading from $CERT_HOSTNAME_FILE"
+			CERT_HOSTNAME=$(cat $CERT_HOSTNAME_FILE)
+		fi
 
-	if [[ ! -z "${CERT_PASSWORD_FILE:-}" ]]
-	then
-		echo "CERT_PASSWORD_FILE is set so reading from $CERT_PASSWORD_FILE"
-		CERT_PASSWORD=$(cat $CERT_PASSWORD_FILE)
-	fi
+		if [[ -z "${CERT_HOSTNAME:-}" ]]; then
+			echo "CERT_HOSTNAME must be set"
+			exit 1
+		fi
 
-	if [[ -z "${CERT_PASSWORD:-}" ]]; then
-		echo "CERT_PASSWORD must be set"
-    	exit 1
-	fi
+		if [[ ! -z "${CERT_PASSWORD_FILE:-}" ]]
+		then
+			echo "CERT_PASSWORD_FILE is set so reading from $CERT_PASSWORD_FILE"
+			CERT_PASSWORD=$(cat $CERT_PASSWORD_FILE)
+		fi
 
-	/bin/bash /opt/healthcatalyst/setupca.sh \
-		&& /bin/bash /opt/healthcatalyst/generateservercert.sh \
-		&& /bin/bash /opt/healthcatalyst/generateclientcert.sh fabricrabbitmquser \
-		&& mkdir -p /opt/certs/testca \
-		&& cp /opt/healthcatalyst/testca/cacert.pem /opt/certs/testca \
-		&& mkdir -p /opt/certs/server \
-		&& cp /opt/healthcatalyst/server/cert.pem /opt/certs/server/ \
-		&& cp /opt/healthcatalyst/server/key.pem /opt/certs/server/ \
-		&& mkdir -p /opt/certs/client \
-		&& cp /opt/healthcatalyst/client/*.p12 /opt/certs/client/
-		
-else
-	echo "certificates already exist so we're not regenerating them"
+		if [[ -z "${CERT_PASSWORD:-}" ]]; then
+			echo "CERT_PASSWORD must be set"
+			exit 1
+		fi
+
+		/bin/bash /opt/healthcatalyst/setupca.sh \
+			&& /bin/bash /opt/healthcatalyst/generateservercert.sh \
+			&& /bin/bash /opt/healthcatalyst/generateclientcert.sh fabricrabbitmquser \
+			&& mkdir -p /opt/certs/testca \
+			&& cp /opt/healthcatalyst/testca/cacert.pem /opt/certs/testca \
+			&& mkdir -p /opt/certs/server \
+			&& cp /opt/healthcatalyst/server/cert.pem /opt/certs/server/ \
+			&& cp /opt/healthcatalyst/server/key.pem /opt/certs/server/ \
+			&& mkdir -p /opt/certs/client \
+			&& cp /opt/healthcatalyst/client/*.p12 /opt/certs/client/
+	else
+		echo "certificates already exist so we're not regenerating them"
+	fi
 fi
 
 MyHostName="${CERT_HOSTNAME:-$(hostname)}"
 
 # copy only the client certs to the web server folder for download
 mkdir -p /app/public/client/ \
+	&& mkdir -p /opt/certs/client \
 	&& cp /opt/certs/client/*.p12 /app/public/client/
 
 echo "you can download the client certificate from this url"
